refactor animation structure again
