+++ LuokeCollection/main/battle/skill_effect.py
@@ -1,0 +1,7 @@
+class SkillEffect:
+    def __init__(self, animator):
+        self.turns = 0
+        self.anim = animator
+
+    def resolve_effect(self):
+        pass
